third attempt

+++ src/main/java/com/driver/Main.java
@@ -8,7 +8,7 @@ public class Main {
         public int product(int x,int y,int z){
             return x*y*z;
         }
-        public int product(double x,double y){
+        public double product(double x,double y){
             return x*y;
         }
     }
